refactor(compression): remove usage of uint16_t for arithmetic
Doing arithmetic with types smaller than 32/64 bit is slower on modern processors
due to partial register writes. This also artificially limits ACL to 65535 samples, segments, etc.

+++ includes/acl/compression/impl/track_stream.h
@@ -316,8 +316,8 @@ namespace acl
 		struct BoneStreams
 		{
 			SegmentContext* segment;
-			uint16_t bone_index;
-			uint16_t parent_bone_index;
+			uint32_t bone_index;
+			uint32_t parent_bone_index;
 			uint32_t output_index;
 
 			RotationTrackStream rotations;
